fixed custom player stats
(cherry picked from commit a3b5e269b93651d76e4246ca93957a1499d55a27)

+++ src/server/game/Entities/Player/Player.h
@@ -289,8 +289,8 @@ typedef std::list<PlayerCreateInfoItem> PlayerCreateInfoItems;
 struct PlayerClassLevelInfo
 {
     PlayerClassLevelInfo()  = default;
-    uint16 basehealth{0};
-    uint16 basemana{0};
+    uint32 basehealth{0};
+    uint32 basemana{0};
 };
 
 struct PlayerClassInfo
@@ -304,7 +304,7 @@ struct PlayerLevelInfo
 {
     PlayerLevelInfo() { for (unsigned char & stat : stats) stat = 0; }
 
-    uint8 stats[MAX_STATS];
+    uint16 stats[MAX_STATS];
 };
 
 typedef std::list<uint32> PlayerCreateInfoSpells;
